@@ -2,77 +2,59 @@
   "name": "curriculum-react",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "start.js",
   "scripts": {
     "start": "cross-env NODE_ENV=development node start.js",
-    "build": "cross-env NODE_ENV=production webpack -p",
-    "test": "jest",
-    "test:coverage": "jest --coverage",
-    "serve:build": "serve -s build -l 4000"
+    "start:ts": "cross-env SKILLERWHALE_LANG=ts npm start",
+    "start:train": "cross-env webpack-dev-server",
+    "start:train:ts": "cross-env SKILLERWHALE_LANG=ts npm run start:train",
+    "build": "cross-env NODE_ENV=production webpack --mode production",
+    "test": "jest /src",
+    "test:ts": "cross-env SKILLERWHALE_LANG=ts jest /src_tsx",
+    "test:coverage": "npm test -- --coverage",
+    "test:ts:coverage": "npm run test:ts -- --coverage",
+    "serve:build": "serve -s build -l 4000",
+    "lint": "eslint . --ext ts,tsx,js,jsx --fix",
+    "compile": "tsc --jsx preserve && npm run lint"
   },
   "keywords": [],
   "author": "",
   "prettier": {
     "semi": false
   },
-  "jest": {
-    "testURL": "http://localhost/",
-    "setupTestFrameworkScriptFile": "./setupTests.js",
-    "moduleFileExtensions": [
-      "js",
-      "jsx"
-    ],
-    "moduleDirectories": [
-      "node_modules"
-    ],
-    "moduleNameMapper": {
-      "\\.(css|styl|less|sass|scss)$": "identity-obj-proxy"
-    },
-    "transform": {
-      "^.+\\.js$": "babel-jest",
-      "^.+\\.jsx$": "babel-jest",
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/src/testing/__mock__/fileTransformer.js"
-    },
-    "collectCoverageFrom": [
-      "src/testing/drive-safe-app/**/*.{js,jsx}",
-      "!src/testing/drive-safe-app/api.js",
-      "!**/node_modules/**"
-    ],
-    "coverageReporters": [
-      "json",
-      "lcovonly",
-      "text",
-      "clover"
-    ]
-  },
   "devDependencies": {
-    "@sheerun/mutationobserver-shim": "^0.3.3",
-    "@testing-library/react": "^10.4.3",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^22.4.4",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "cross-env": "^5.2.1",
-    "css-loader": "^0.28.11",
-    "file-loader": "^1.1.11",
-    "html-webpack-plugin": "^3.2.0",
+    "@babel/core": "^7.18.2",
+    "@babel/plugin-syntax-jsx": "^7.17.12",
+    "@babel/plugin-transform-react-jsx": "^7.17.12",
+    "@babel/preset-env": "^7.18.2",
+    "@faker-js/faker": "^7.1.0",
+    "@testing-library/react": "^13.3.0",
+    "@types/jest": "^28.1.1",
+    "@typescript-eslint/parser": "^5.27.0",
+    "babel-loader": "^8.2.5",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.17.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "html-webpack-plugin": "^5.5.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^22.4.4",
-    "serve": "^11.3.2",
-    "webpack": "^4.39.3",
-    "webpack-cli": "^3.3.8",
-    "webpack-dev-server": "^3.8.0"
+    "jest": "^28.1.0",
+    "prettier": "^2.6.2",
+    "serve": "^13.0.2",
+    "ts-jest": "^28.0.4",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.7.3",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.1"
   },
   "dependencies": {
-    "lodash": "^4.17.21",
-    "prop-types": "^15.7.2",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "skiller-whale-sync": "^1.6.0"
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "skiller-whale-sync": "^1.7.0"
   }
 }
