@@ -1,17 +1,12 @@
-import React from "react"
-import ReactDOM from "react-dom"
-
+import { createRoot } from "react-dom/client"
 import App from "./App"
 // import App from "./hooks/useState"
 // import App from "./hooks/useEffect"
-// import App from "./refs/refClassComponent"
-// import App from "./refs/refFunctionComponent"
-// import App from "./refs/childRefs"
+// import App from "./refs"
 // import App from "./component-lifecycle"
 // import App from "./testing/drive-safe-app"
 // import App from "./portals"
 // import App from "./suspense"
-
-const root = document.getElementById("root")
-
-ReactDOM.render(<App />, root)
+const container = document.getElementById("root")
+const root = createRoot(container)
+root.render(<App />)
