@@ -1,13 +1,6 @@
-import React from "react"
-
-class App extends React.Component {
-  render() {
-    return (
-      <div className="container">
-        <h1>App</h1>
-      </div>
-    )
-  }
-}
-
+const App = () => (
+  <div className="container">
+    <h1>App</h1>
+  </div>
+)
 export default App
