@@ -1,3 +1,3 @@
 const swsync = require("skiller-whale-sync")
 
-swsync("webpack-dev-server -d")
+swsync(`cross-env webpack-dev-server`)
